refactor: move git_add to shared location
git_add will be reused by many subcommands, so moving it to shared
location.
