+++ src/pipeline_migration/utils.py
@@ -1,6 +1,9 @@
 import hashlib
+import logging
 from dataclasses import dataclass, field
 from typing import Any
+from pathlib import Path
+import subprocess
 
 from pipeline_migration.types import FilePath
 
@@ -15,7 +18,10 @@ __all__ = [
     "is_true",
     "load_yaml",
     "YAMLStyle",
+    "git_add",
 ]
+
+logger = logging.getLogger("utils")
 
 
 def is_true(value: str) -> bool:
@@ -120,3 +126,23 @@ def dump_yaml(yaml_file: FilePath, data: Any, style: YAMLStyle | None = None) ->
 def file_checksum(file_path: FilePath) -> str:
     with open(file_path, "rb") as f:
         return hashlib.sha256(f.read()).hexdigest()
+
+
+def git_add(file_path: FilePath) -> None:
+    """Git add given file
+
+    The git-add command may fail due to any reason, e.g. git command is not available in the system,
+    in which case just logging a message and terminate quietly.
+
+    :param file_path: an absolute path to a file.
+    :type file_path: FilePath
+    :raises ValueError: if given file path is not an absolute path.
+    """
+    fp = Path(file_path)
+    if not fp.is_absolute():
+        raise ValueError(f"File path {file_path} is not an absolute path.")
+    cmd = ["git", "add", fp.name]
+    try:
+        subprocess.run(cmd, cwd=fp.parent, capture_output=True, check=True)
+    except subprocess.CalledProcessError as e:
+        logger.warning("%s is not added to git index: %s", file_path, e.stderr)
